requirements.txt
fastapi
uvicorn
sqlalchemy
line-bot-sdk
python-dotenv
psycopg2-binary
asyncpg
GeoAlchemy2
